Applying SOLID: Added abstract layer between controller and concrete service classes.
