+++ src/main/java/com/makeprojects/library/elib/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.makeprojects.library.elib.controller;
 
+import com.makeprojects.library.elib.core.service.definition.MemberService;
 import com.makeprojects.library.elib.entity.Member;
-import com.makeprojects.library.elib.service.MemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -24,25 +24,25 @@ public class MemberController {
     // Moved to AUthController
 //    @PostMapping("/add")
 //    public ResponseEntity<Member> addMember(@RequestBody Member member) {
-//        Member addedMember = this.memberService.addMember(member);
+//        Member addedMember = this.memberService.create(member);
 //        return new ResponseEntity<>(addedMember, HttpStatus.CREATED);
 //    }
 
     @PostMapping("/update")
     public ResponseEntity<Member> updateMember(@RequestBody Member member) {
-        Member updatedMember = this.memberService.updateMember(member);
+        Member updatedMember = this.memberService.update(member);
         return new ResponseEntity<>(updatedMember, HttpStatus.OK);
     }
 
     @GetMapping("/list")
     public ResponseEntity<List<Member>> getAllMembers() {
-        List<Member> members = this.memberService.getAllMembers();
+        List<Member> members = this.memberService.getAll();
         return new ResponseEntity<>(members, HttpStatus.OK);
     }
 
     @GetMapping("/{memberId}")
     public ResponseEntity<Member> getMemberById(@PathVariable UUID memberId) {
-        Member member = this.memberService.getMemberById(memberId);
+        Member member = this.memberService.get(memberId);
         return new ResponseEntity<>(member, HttpStatus.OK);
     }
