+++ src/main/java/com/makeprojects/library/elib/core/service/definition/ArithmeticService.java
@@ -1,0 +1,8 @@
+package com.makeprojects.library.elib.core.service.definition;
+
+import com.makeprojects.ewallet.shared.core.definition.Service;
+
+public interface ArithmeticService extends Service {
+
+    double divideTwoNumbers(int a, int b);
+}
