+++ src/main/java/com/makeprojects/library/elib/core/service/definition/BookService.java
@@ -1,0 +1,7 @@
+package com.makeprojects.library.elib.core.service.definition;
+
+import com.makeprojects.ewallet.shared.core.definition.CRUDService;
+import com.makeprojects.library.elib.entity.Book;
+
+public interface BookService extends CRUDService<Book> {
+}
