--- src/main/java/com/makeprojects/library/elib/service/MyService.java
@@ -1,11 +1,0 @@
-package com.makeprojects.library.elib.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class MyService {
-
-    public double divideTwoNumbers(int a, int b) {
-        return (double) a/b;
-    }
-}
