+++ src/test/java/com/makeprojects/library/elib/service/ArithmeticServiceTest.java
@@ -1,18 +1,19 @@
 package com.makeprojects.library.elib.service;
 
 import com.makeprojects.library.elib.ELibraryApplication;
+import com.makeprojects.library.elib.core.service.definition.ArithmeticService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest(classes = ELibraryApplication.class)
-public class MyServiceTest {
+public class ArithmeticServiceTest {
 
-    private MyService myService;
+    private ArithmeticService myService;
 
     @Autowired
-    public MyServiceTest(MyService myService) {
+    public ArithmeticServiceTest(ArithmeticService myService) {
         this.myService = myService;
     }
